requirements.txt
# Python dependencies for Resume Analyzer & Job Match Prediction System

# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0
gunicorn>=21.2.0

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Document Processing & PDF Generation
PyPDF2>=3.0.0
python-docx>=0.8.11
reportlab>=4.0.0

# AI APIs for Resume Parsing
openai>=1.0.0           # For OpenAI GPT models
google-generativeai>=0.3.0  # For Google Gemini
anthropic>=0.7.0        # For Anthropic Claude (optional)

# Environment Variables
python-dotenv>=1.0.0

# Excel Report Generation
openpyxl>=3.1.0
XlsxWriter>=3.1.0

# Natural Language Processing (Optional)
# nltk>=3.8.0
# spacy>=3.6.0

# Web Framework (for future API/Web interface)
# flask>=3.0.0
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Progress Bars
tqdm>=4.66.0

# Utilities
joblib>=1.3.0
